requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
et-xmlfile==1.1.0
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
multidict==6.0.5
openpyxl==3.1.2
pyaes==1.6.1
pyasn1==0.5.1
python-dotenv==1.0.1
rsa==4.9
SQLAlchemy==2.0.27
Telethon==1.34.0
typing_extensions==4.10.0
yarl==1.9.4
